wiki_scrapper/phase2/categorizer/requirements.txt
beautifulsoup4==4.13.3
mysql-connector-python==9.0.0
pyspark==3.5.4
